requirements.txt
aniso8601==1.3.0
argh==0.26.2
asn1crypto==0.23.0
blinker==1.4
brotlipy==0.6.0
certifi==2017.7.27.1
cffi==1.11.2
chardet==3.0.4
click==6.7
construct==2.8.16
cryptography==1.8.2
cssutils==1.0.2
Django==1.11.5
djangorestframework==3.6.4
EditorConfig==0.12.1
Flask==0.12.2
Flask-JWT==0.3.2
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.3.2
h2==2.6.2
hpack==3.0.0
html2text==2016.9.19
hyperframe==4.0.2
idna==2.6
itsdangerous==0.24
Jinja2==2.10
jsbeautifier==1.6.14
kaitaistruct==0.6
MarkupSafe==1.0
mitmproxy==2.0.2
packaging==16.8
passlib==1.7.1
pathtools==0.1.2
pyasn1==0.2.3
pycparser==2.18
pydivert==2.0.9
Pygments==2.2.0
PyJWT==1.4.2
pyOpenSSL==16.2.0
pyparsing==2.2.0
pyperclip==1.5.32
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
requests==2.18.4
ruamel.yaml==0.13.14
six==1.11.0
sortedcontainers==1.5.7
SQLAlchemy==1.1.15
tornado==4.4.3
urllib3==1.22
urwid==1.3.1
watchdog==0.8.3
Werkzeug==0.12.2
